+++ .vscode/launch.json
@@ -1,0 +1,11 @@
+{
+    "version": "0.2.0",
+    "configurations": [
+        
+        {
+            "name": "Flutter mobile app",
+            "type": "flutter",
+            "request": "launch"
+        }
+    ]
+}
